객체 참조 오류 해결 - refresh_token-member
